+++ example/wxpub_oauth.py
@@ -30,10 +30,15 @@ print(jsapi_ticket)
 # -------
 # 获取微信小程序open_id
 '''
-小程序调用wx.login() 获取 临时登录凭证code ，并回传到开发者服务器。
+小程序调用 wx.login() 获取 临时登录凭证 code ，并回传到开发者服务器。
 临时登录凭证code只能使用一次
 '''
 
+# 微信小程序的 App ID
+wx_lite_app_id = 'wx9283883h4773765'
+# 微信小程序的 App Secret
+wx_lite_app_secret = 'da38db82d0e495a037aec19df178572c'
 wx_lite_code = '011TFt8z15UDqh0xcF5z1A5E8z1TFt8W'
-wx_lite_openid = pingpp.WxpubOauth.get_openid(wx_app_id, wx_app_secret, wx_lite_code)
+wx_lite_openid = pingpp.WxpubOauth.get_lite_openid(
+    wx_lite_app_id, wx_lite_app_secret, wx_lite_code)
 print(wx_lite_openid)